*: move reference pkg under docker
